change type of attributes BillData

+++ src/Types/BillPayments/BillData.php
@@ -6,10 +6,10 @@ use WeDevBr\Celcoin\Types\Data;
 
 class BillData extends Data
 {
-    public int $value;
-    public int $originalValue;
-    public int $valueWithDiscount;
-    public int $valueWithAdditional;
+    public float $value;
+    public float $originalValue;
+    public float $valueWithDiscount;
+    public float $valueWithAdditional;
 
     public function __construct(array $data = [])
     {
